Moved requirements
